+++ packages/app-store/googlecalendar/lib/CalendarService.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import { calendar_v3 } from "@googleapis/calendar";
 import type { Prisma } from "@prisma/client";
 import { OAuth2Client } from "googleapis-common";
-import { calendar_v3 } from "@googleapis/calendar";
 import { RRule } from "rrule";
 import { v4 as uuid } from "uuid";
 
@@ -20,7 +20,7 @@ import {
 import { formatCalEvent } from "@calcom/lib/formatCalendarEvent";
 import logger from "@calcom/lib/logger";
 import { safeStringify } from "@calcom/lib/safeStringify";
-import { GoogleService } from "@calcom/lib/server/service/google";
+import { SelectedCalendarRepository } from "@calcom/lib/server/repository/selectedCalendar";
 import prisma from "@calcom/prisma";
 import type {
   Calendar,
@@ -539,12 +539,12 @@ export default class GoogleCalendarService implements Calendar {
       // Only calendars of other integrations selected
       return [];
     }
-    async function getCalIds() {
+    const getCalIds = async () => {
       if (selectedCalendarIds.length !== 0) return selectedCalendarIds;
-      const cals = await GoogleService.getAllCalendars(calendar, ["id"]);
+      const cals = await this.getAllCalendars(calendar, ["id"]);
       if (!cals.length) return [];
       return cals.reduce((c, cal) => (cal.id ? [...c, cal.id] : c), [] as string[]);
-    }
+    };
 
     try {
       const calsIds = await getCalIds();
@@ -605,7 +605,7 @@ export default class GoogleCalendarService implements Calendar {
             status: 200,
             statusText: "OK",
             data: {
-              items: await GoogleService.getAllCalendars(calendar),
+              items: await this.getAllCalendars(calendar),
             },
           })
       );
@@ -650,10 +650,9 @@ export default class GoogleCalendarService implements Calendar {
       },
     });
     const response = res.data;
-    await GoogleService.upsertSelectedCalendar({
+    await this.upsertSelectedCalendar({
       userId: this.credential.userId!,
       externalId: calendarId,
-      credentialId: this.credential.id,
       googleChannelId: response?.id,
       googleChannelKind: response?.kind,
       googleChannelResourceId: response?.resourceId,
@@ -684,10 +683,9 @@ export default class GoogleCalendarService implements Calendar {
       .catch((err) => {
         console.warn(JSON.stringify(err));
       });
-    await GoogleService.upsertSelectedCalendar({
+    await this.upsertSelectedCalendar({
       userId: this.credential.userId!,
       externalId: calendarId,
-      credentialId: this.credential.id,
       googleChannelId: null,
       googleChannelKind: null,
       googleChannelResourceId: null,
@@ -712,6 +710,89 @@ export default class GoogleCalendarService implements Calendar {
     };
     const data = await this.fetchAvailability(parsedArgs);
     await this.setAvailabilityInCache(parsedArgs, data);
+  }
+
+  async createSelectedCalendar(
+    data: Omit<Prisma.SelectedCalendarUncheckedCreateInput, "integration" | "credentialId">
+  ) {
+    return await SelectedCalendarRepository.create({
+      ...data,
+      integration: this.integrationName,
+      credentialId: this.credential.id,
+    });
+  }
+
+  async upsertSelectedCalendar(
+    data: Omit<Prisma.SelectedCalendarUncheckedCreateInput, "integration" | "credentialId">
+  ) {
+    return await SelectedCalendarRepository.upsert({
+      ...data,
+      integration: this.integrationName,
+      credentialId: this.credential.id,
+    });
+  }
+
+  async getAllCalendars(
+    calendar: calendar_v3.Calendar,
+    fields: string[] = ["id", "summary", "primary", "accessRole"]
+  ): Promise<calendar_v3.Schema$CalendarListEntry[]> {
+    let allCalendars: calendar_v3.Schema$CalendarListEntry[] = [];
+    let pageToken: string | undefined;
+
+    try {
+      do {
+        const response: any = await calendar.calendarList.list({
+          fields: `items(${fields.join(",")}),nextPageToken`,
+          pageToken,
+          maxResults: 250, // 250 is max
+        });
+
+        allCalendars = [...allCalendars, ...(response.data.items ?? [])];
+        pageToken = response.data.nextPageToken;
+      } while (pageToken);
+
+      return allCalendars;
+    } catch (error) {
+      logger.error("Error fetching all Google Calendars", { error });
+      throw error;
+    }
+  }
+
+  async getPrimaryCalendar(
+    calendar: calendar_v3.Calendar,
+    fields: string[] = ["id", "summary", "primary", "accessRole"]
+  ): Promise<calendar_v3.Schema$CalendarListEntry | null> {
+    let pageToken: string | undefined;
+    let firstCalendar: calendar_v3.Schema$CalendarListEntry | undefined;
+
+    try {
+      do {
+        const response: any = await calendar.calendarList.list({
+          fields: `items(${fields.join(",")}),nextPageToken`,
+          pageToken,
+          maxResults: 250, // 250 is max
+        });
+
+        const cals = response.data.items ?? [];
+        const primaryCal = cals.find((cal: calendar_v3.Schema$CalendarListEntry) => cal.primary);
+        if (primaryCal) {
+          return primaryCal;
+        }
+
+        // Store the first calendar in case no primary is found
+        if (cals.length > 0 && !firstCalendar) {
+          firstCalendar = cals[0];
+        }
+
+        pageToken = response.data.nextPageToken;
+      } while (pageToken);
+
+      // should not be reached because Google Cal always has a primary cal
+      return firstCalendar ?? null;
+    } catch (error) {
+      logger.error("Error in `getPrimaryCalendar`", { error });
+      throw error;
+    }
   }
 }
